public/Backend/requirements.txt
fastapi>=0.95
uvicorn[standard]>=0.18
sqlalchemy>=1.4
asyncpg>=0.27
psycopg2-binary>=2.9
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3.0
pydantic>=2.0
python-multipart>=0.0.6
requests>=2.31
typing-extensions>=4.0
email-validator>=1.3.1